Explain 'features/new-model-structure's purpose
There where a few problems with how the `model`s where used in
combination with `powerplant`s.
It was intended that for each new `powerplant` instance an accompanying
new `model` instance was to be created but neither the documentation nor
the interface made this fact particularly clear so users started to
reuse the same `model` instance for different `powerplant`s. This leads
to various problems, issue #7 being one of them.
The purpose of this branch is to change the `powerplant`s interface in
such a way, that one no longer supplies a `model` instance to a
`powerplant` but a `model` class instead. The instance will then be
created in the `powerplant` constructor ensuring a fresh `model`
instance for each `powerplant`.