Enable usage of classes as model parameters
This means that users now have the opportunity to use both, a model
instance or a model class as valid values for the model parameter to
powrplant classes. There are a few other changes mixed into this commit
which probably should have been split out into separate ones. Sorry for
that. What follows is a short explanation of the other changes, this
commit comes along with.

As one can now uses classes for the model argument, the default
arguments to the `Photovoltaic` and `WindPowerPlant` classes have been
changed to use this feature. This also dodges bugs created by Python's
"evaluate default arguments on compile time"[^0] behaviour.

There also was duplicate functionality in these classes. It has been
abstracted out into a common `Base` class which also acts as an abstract
base class, forcing anyone who writes a subclass to override the
`feedin` method.

[^0]: See e.g.: http://effbot.org/zone/default-values.htm for an
      explanation.