Move scaling of feedin timeseries
Scaling of feedin timeseries is moved to Photovoltaic and WindPowerPlant
classes as different scaling methods are suitable for those technologies
(e.g. scaling by area can only be used for PV).
Instead of checking for different kwargs using if statements a "scaling"
parameter is introduced to specify if and how to scale. The parameter
"scaling_value" specifies the amount to scale the time series up/down by
and depending on the scaling method could be a total module area or
installed capacity.
Scaling is added to simple_feedin example in order to check if it's
working correctly.