Optimize series query
This formulation is shorter and should be equivalent to the longer one.
Using `distinct` also reduces the amount of data which has to be
transmitted and it reduces the time spent in the `set` constructor.
On my machine, querying the windpowerlib variables for one location and
for 2012 is nearly a minute faster with 4m22s down from 5m17s.