Improve `self.series.values()` format
Instead of being a set of lists, it's now a properly sorted sequence of
`(start, stop, value)` triples and it's properly filtered, even if the
queried `start` and `stop` would've been within a segment.
Please don't get tripped up by the weird formatting. This is just to
keep the diff small and easier to read. It will be fixed in a dedicated
commit.