Fix series query
There's a lot going on here. In short, the query which got the series
values from the database has been refactored. It now actually works
correctly and instead of being split up into multiple small queries, it
is now only one query. The cool thing is, this makes the code easier to
understand, a lot faster and only around five lines longer in total.
To explain the diff in a bit more detail, there where a bunch of
problems with old code. First, the `for q in [` part shows that it
issued one query for every distinct combination of variable name, height
and location, and tried to chain the results together in the
`for series in chain(` part. Doing that many isolated queries was
obviously quite slow. Doing it this way also meant that we had to issue
a query getting the heights levels of each variable and intersect it
with the selected heights in a pretty complicated way. On top of that,
the query used the wrong joins, leading huge and incorrect results.
The new code still has to use separate queries to get the measurement
locations closest to the selected points and within the selected
regions, but after that, there's only one query getting all the
remaining data. This new query gets the correct results by joining along
the defined relationships and it is also more concise and readable by
filtering using the `in_` operator, which makes filtering selected
locations, heights and variables a lot easier. The only side effect was,
that it was convenient to drop the `"all"` special argument and use
`None` instead, as `.filter(None)` has the effect of not filtering at
all. This is also reflected in the docstrings. Last but not least,
querying the database in this way makes it really easy to sort the
result and then use `groupby` using a similar key to get correctly
ordered (time segment,value) combinations for each (location, variable,
height) combination.