Fix list comprehension indentation
I left it one level indented to far in order to keep the diff as small
as possible. Now it's time to de-dent it. The only other change is that
`series.timespan.segments` got changed to `timespan.segments` to save a
line. But since `series.timespan` and `timespan` should be identical
objects, this should not change the semantics of the code.