Fix `Weather` docstring example
The example had to be adjusted for the `heights` parameter. I also took
the liberty of making the queried time period a lot smaller and using
`"pvlib"` instead of `["T"]` to select variables. That way the example
can immediately be followed with retrieving a Pvlib `DataFrame` which,
thanks to the smaller query window, is small enough to be printed for a
visual sanity check.