Tentatively fix "windpowerlib" `MultiIndex`
The problem was that while preventing the "pvlib" columns from being a
`MultiIndex` I accidentally also prevented the "windpowerlib" columns
from being a `MultiIndex`. But for the "windpowerlib" we want a
`MultiIndex` with the variable name on the first level and the height on
the second level. This commit hopefully fixes that. I can't test it
though, as only one test fails, but for a different reason.