Adjust moved code to expected types
The code previously worked on the `series` dictionary as a whole, but
the type annotation and the documentation hinted at the fact that
`deduplicate` only works on one list, i.e. on dictionary value, at a
time. Doing so makes the code more compact and doesn't complicate the
`deduplicate` call too much.