Allow querying multiple time intervals
This is for cases where one wants to query weather data for multiple,
discontinuous time intervals. Supporting this means one doesn't have to
do multiple queries for these cases.
This is, aside from conversion to `pandas.Timestamp`s, a generalization
of the `start` and `stop` parameters, so these are optional now.